fix!: writing to same output as input
Putting the same file as input and output made the file end up blank,
this seems to be normal behaviour on bash and the terminal works, so the
cli was fixed to reflect more transparently that behaviour. Now the
input is always the last parameter. And for convinience, a write option
was made to write to the same file as the input.